refactor : SocketHandler 함수 분리
