+++ backend/routes/api/userList.ts
@@ -8,7 +8,6 @@ router.get('/', (req, res, next) => {
   const roomTitle = `${req.query.roomTitle}`;
   const targetRoom = utils.findRoomOnTitle(socketData, roomTitle);
   const userList = targetRoom?.userId.filter((val, idx, arr) => arr.indexOf(val) === idx);
-  console.log("userList", userList);
   if (!targetRoom) res.json({ list: ['bad'] });
 
   res.json({ list: userList });
