--- backend/sessions/p-sQzXaONBl8KD_yaQijRxJC_4oPP_tX.json
@@ -1,1 +1,0 @@
-{"cookie":{"originalMaxAge":null,"expires":null,"secure":false,"httpOnly":true,"path":"/"},"__lastAccess":1637136863452,"isEnterd":true}
