--- backend/sessions/rQ333ZQlo1ci-MTGVREs4EMYCNpXS1Xu.json
@@ -1,1 +1,0 @@
-{"cookie":{"originalMaxAge":null,"expires":null,"secure":false,"httpOnly":true,"path":"/"},"__lastAccess":1637134654252,"isEnterd":true}
