+++ frontend/src/context/MyContext.tsx
@@ -1,11 +1,11 @@
-import { createContext, useContext } from 'react';
-import io from 'socket.io-client';
+import { createContext, ReactElement, useContext } from 'react';
+import io, { Socket } from 'socket.io-client';
 import config from '../config.host.json';
 
-const socket: any = io(`${config.localhost}/music`);
-const SocketContext = createContext(null);
+const socket: Socket = io(`${config.localhost}/music`);
+const SocketContext = createContext(socket);
 
-const ContextProvider = ({ children }: { children: any }) => {
+const ContextProvider = ({ children }: { children: ReactElement }) => {
   return <SocketContext.Provider value={socket}>{children}</SocketContext.Provider>;
 };
